refactor: reorganize communication module into notification module
