+++ src/apps/backend/modules/notification/types.py
